+++ src/types/goal.ts
@@ -1,0 +1,5 @@
+export type GoalSettings = {
+  currentNetWorth: number
+  targetNetWorth: number
+  monthlyContribution: number
+}